pubsub/requirements.txt
Flask==2.0.1
pytest==5.3.2; python_version > "3.0"
pytest==5.3.2; python_version < "3.0"
gunicorn==20.1.0
google-cloud-pubsub==1.7.0